chore: Refactor organise token-expressions
