+++ journeyapps-evaluator/src/definitions/ObjectRefInterface.ts
