+++ journeyapps-evaluator/src/definitions/TypeInterface.ts
