+++ journeyapps-evaluator/src/index.ts
@@ -1,16 +1,9 @@
-export * from './token-expressions/ConstantTokenExpression';
-export * from './token-expressions/FormatShorthandTokenExpression';
-export * from './token-expressions/FunctionTokenExpression';
-export * from './token-expressions/LegacyFunctionTokenExpression';
-export * from './token-expressions/PrimitiveConstantTokenExpression';
-export * from './token-expressions/ShorthandTokenExpression';
-export * from './token-expressions/TokenExpression';
-export * from './token-expressions/FormatString';
+export * from './token-expressions';
 
-export * from './types/TypeInterface';
-export * from './types/ObjectRefInterface';
+export * from './definitions/TypeInterface';
+export * from './definitions/ObjectRefInterface';
+export * from './definitions/FormatStringScope';
 
-export * from './scope/FormatStringScope';
 export * from './scope/VariableFormatStringScope';
 
 export * from './tools';
