+++ journeyapps-evaluator/tsconfig.json
@@ -5,6 +5,7 @@
     "declaration": true,
     "composite": true,
     "outDir": "dist",
+    "rootDir": "src",
     "tsBuildInfoFile": "dist/tsconfig.tsbuildinfo"
   },
   "include": ["./src/**/*.ts"],
